Refactor navigation out of AddObsModal
